+++ hibernate-orm-panache-resteasy/src/main/java/org/acme/hibernate/orm/panache/Fruit.java
@@ -1,4 +1,4 @@
-package com.acme.hibernate.orm.panache;
+package org.acme.hibernate.orm.panache;
 
 import javax.persistence.Cacheable;
 import javax.persistence.Column;
