+++ hibernate-orm-panache-resteasy/src/main/java/org/acme/hibernate/orm/panache/FruitResource.java
@@ -1,9 +1,9 @@
-package com.acme.hibernate.orm;
+package org.acme.hibernate.orm.panache;
+
+import java.util.List;
 
 import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
 import javax.json.Json;
-import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -19,25 +19,23 @@ import javax.ws.rs.ext.Provider;
 
 import org.jboss.resteasy.annotations.jaxrs.PathParam;
 
+import io.quarkus.panache.common.Sort;
+
 @Path("fruits")
 @ApplicationScoped
 @Produces("application/json")
 @Consumes("application/json")
 public class FruitResource {
 
-    @Inject
-    EntityManager entityManager;
-
     @GET
-    public Fruit[] get() {
-        return entityManager.createNamedQuery("Fruits.findAll", Fruit.class)
-              .getResultList().toArray(new Fruit[0]);
+    public List<Fruit> get() {
+        return Fruit.listAll(Sort.by("name"));
     }
 
     @GET
     @Path("{id}")
-    public Fruit getSingle(@PathParam Integer id) {
-        Fruit entity = entityManager.find(Fruit.class, id);
+    public Fruit getSingle(@PathParam Long id) {
+        Fruit entity = Fruit.findById(id);
         if (entity == null) {
             throw new WebApplicationException("Fruit with id of " + id + " does not exist.", 404);
         }
@@ -47,29 +45,29 @@ public class FruitResource {
     @POST
     @Transactional
     public Response create(Fruit fruit) {
-        if (fruit.getId() != null) {
+        if (fruit.id != null) {
             throw new WebApplicationException("Id was invalidly set on request.", 422);
         }
 
-        entityManager.persist(fruit);
+        fruit.persist();
         return Response.ok(fruit).status(201).build();
     }
 
     @PUT
     @Path("{id}")
     @Transactional
-    public Fruit update(@PathParam Integer id, Fruit fruit) {
-        if (fruit.getName() == null) {
+    public Fruit update(@PathParam Long id, Fruit fruit) {
+        if (fruit.name == null) {
             throw new WebApplicationException("Fruit Name was not set on request.", 422);
         }
 
-        Fruit entity = entityManager.find(Fruit.class, id);
+        Fruit entity = Fruit.findById(id);
 
         if (entity == null) {
             throw new WebApplicationException("Fruit with id of " + id + " does not exist.", 404);
         }
 
-        entity.setName(fruit.getName());
+        entity.name = fruit.name;
 
         return entity;
     }
@@ -77,12 +75,12 @@ public class FruitResource {
     @DELETE
     @Path("{id}")
     @Transactional
-    public Response delete(@PathParam Integer id) {
-        Fruit entity = entityManager.getReference(Fruit.class, id);
+    public Response delete(@PathParam Long id) {
+        Fruit entity = Fruit.findById(id);
         if (entity == null) {
             throw new WebApplicationException("Fruit with id of " + id + " does not exist.", 404);
         }
-        entityManager.remove(entity);
+        entity.delete();
         return Response.status(204).build();
     }
