+++ hibernate-orm-resteasy/src/main/java/org/acme/hibernate/orm/FruitResource.java
@@ -1,9 +1,9 @@
-package com.acme.hibernate.orm.panache;
-
-import java.util.List;
+package org.acme.hibernate.orm;
 
 import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
 import javax.json.Json;
+import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -19,23 +19,25 @@ import javax.ws.rs.ext.Provider;
 
 import org.jboss.resteasy.annotations.jaxrs.PathParam;
 
-import io.quarkus.panache.common.Sort;
-
 @Path("fruits")
 @ApplicationScoped
 @Produces("application/json")
 @Consumes("application/json")
 public class FruitResource {
 
+    @Inject
+    EntityManager entityManager;
+
     @GET
-    public List<Fruit> get() {
-        return Fruit.listAll(Sort.by("name"));
+    public Fruit[] get() {
+        return entityManager.createNamedQuery("Fruits.findAll", Fruit.class)
+              .getResultList().toArray(new Fruit[0]);
     }
 
     @GET
     @Path("{id}")
-    public Fruit getSingle(@PathParam Long id) {
-        Fruit entity = Fruit.findById(id);
+    public Fruit getSingle(@PathParam Integer id) {
+        Fruit entity = entityManager.find(Fruit.class, id);
         if (entity == null) {
             throw new WebApplicationException("Fruit with id of " + id + " does not exist.", 404);
         }
@@ -45,29 +47,29 @@ public class FruitResource {
     @POST
     @Transactional
     public Response create(Fruit fruit) {
-        if (fruit.id != null) {
+        if (fruit.getId() != null) {
             throw new WebApplicationException("Id was invalidly set on request.", 422);
         }
 
-        fruit.persist();
+        entityManager.persist(fruit);
         return Response.ok(fruit).status(201).build();
     }
 
     @PUT
     @Path("{id}")
     @Transactional
-    public Fruit update(@PathParam Long id, Fruit fruit) {
-        if (fruit.name == null) {
+    public Fruit update(@PathParam Integer id, Fruit fruit) {
+        if (fruit.getName() == null) {
             throw new WebApplicationException("Fruit Name was not set on request.", 422);
         }
 
-        Fruit entity = Fruit.findById(id);
+        Fruit entity = entityManager.find(Fruit.class, id);
 
         if (entity == null) {
             throw new WebApplicationException("Fruit with id of " + id + " does not exist.", 404);
         }
 
-        entity.name = fruit.name;
+        entity.setName(fruit.getName());
 
         return entity;
     }
@@ -75,12 +77,12 @@ public class FruitResource {
     @DELETE
     @Path("{id}")
     @Transactional
-    public Response delete(@PathParam Long id) {
-        Fruit entity = Fruit.findById(id);
+    public Response delete(@PathParam Integer id) {
+        Fruit entity = entityManager.getReference(Fruit.class, id);
         if (entity == null) {
             throw new WebApplicationException("Fruit with id of " + id + " does not exist.", 404);
         }
-        entity.delete();
+        entityManager.remove(entity);
         return Response.status(204).build();
     }
